Refactoring: Google auth button as separate react component
